requirements.txt
dnspython==2.6.1
PyGithub==2.4.0
PyGithub==2.4.0
python-dotenv==1.0.1
python_whois==0.9.4
Requests==2.32.3
tqdm==4.66.5